+++ rubato/utils/__init__.py
@@ -3,6 +3,7 @@ This module houses low-dependency utilities widely used in rubato.
 """
 from .path import *
 from .error import *
+from .rb_time import DelayedTask, FramesTask, ScheduledTask, Time
 from .rb_math import Math
 from .noise import Noise
 from .vector import Vector
@@ -15,6 +16,5 @@ from .events import *
 from .radio import Radio, Listener
 from .camera import Camera
 from .draw import Draw, DrawTask
-from .rb_time import DelayedTask, FramesTask, ScheduledTask, Time
 from .sound import Sound
 from .debug import Debug